Move error handling to dedicated submodule
