+++ python/spglib/__init__.py
@@ -8,6 +8,7 @@
 from ._internal import *  # noqa: F403
 from ._version import __version__, __version_tuple__  # noqa: F401
 from .cell import *  # noqa: F403
+from .error import *  # noqa: F403
 from .kpoints import *  # noqa: F403
 from .msg import *  # noqa: F403
 from .reduce import *  # noqa: F403
@@ -15,7 +16,6 @@ from .spg import *  # noqa: F403
 
 # fmt: off
 from .spglib import (  # noqa: F401
-    get_error_message,
     get_hall_number_from_symmetry,
     get_version,
     spg_get_commit,
