+++ python/spglib/_internal.py
@@ -9,12 +9,9 @@ import numpy as np
 from numpy._typing import ArrayLike
 
 from . import _spglib
+from .error import _set_no_error, _set_or_throw_error
 from .spg import SpgCell, SpglibDataset
-from .spglib import (
-    _expand_cell,
-    _set_no_error,
-    _set_or_throw_error,
-)
+from .spglib import _expand_cell
 
 __all__ = [
     "get_pointgroup",
