+++ python/spglib/cell.py
@@ -8,8 +8,9 @@ from __future__ import annotations
 import numpy as np
 
 from . import _spglib
+from .error import _set_no_error, _set_or_throw_error
 from .spg import SpgCell
-from .spglib import _expand_cell, _set_no_error, _set_or_throw_error
+from .spglib import _expand_cell
 
 __all__ = [
     "standardize_cell",
