+++ python/spglib/msg.py
@@ -13,6 +13,7 @@ from numpy.typing import ArrayLike, NDArray
 
 from . import _spglib
 from ._compat.typing import TypeAlias
+from .error import _set_no_error, _set_or_throw_error
 from .spglib import (
     DictInterface,
     Lattice,
@@ -20,8 +21,6 @@ from .spglib import (
     Numbers,
     Positions,
     _expand_cell,
-    _set_no_error,
-    _set_or_throw_error,
 )
 
 __all__ = [
